Quick clean-up of _impl::crib
- move crib-specific functions out of type.h
- flattened namespaces
