+++ include/cnl/_impl/fixed_point/type.h
@@ -41,35 +41,6 @@ namespace cnl {
         struct is_fixed_point<fixed_point<Rep, Exponent, Radix>>
                 : public std::true_type {
         };
-
-        namespace fp {
-            ////////////////////////////////////////////////////////////////////////////////
-            // cnl::_impl::float_of_size
-
-            template<int NumBits, class Enable = void>
-            struct float_of_size;
-
-            template<int NumBits>
-            struct float_of_size<NumBits, enable_if_t<NumBits <= sizeof(float)*CHAR_BIT>> {
-                using type = float;
-            };
-
-            template<int NumBits>
-            struct float_of_size<NumBits, enable_if_t<sizeof(float)*CHAR_BIT < NumBits && NumBits <= sizeof(double)*CHAR_BIT>> {
-                using type = double;
-            };
-
-            template<int NumBits>
-            struct float_of_size<NumBits, enable_if_t<sizeof(double)*CHAR_BIT < NumBits && NumBits <= sizeof(long double)*CHAR_BIT>> {
-                using type = long double;
-            };
-
-            ////////////////////////////////////////////////////////////////////////////////
-            // cnl::_impl::float_of_same_size
-
-            template<class T>
-            using float_of_same_size = typename float_of_size<_impl::width<T>::value>::type;
-        }
     }
 
     /// \brief literal real number approximation that uses fixed-point arithmetic
